packages/alma-connectors/pyproject.toml
[project]
name = "alma-connectors"
version = "0.1.0"
description = "Source adapters for Alma Atlas (BigQuery, Snowflake, Postgres, dbt)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alma-ports",
]

[project.optional-dependencies]
bigquery = ["google-cloud-bigquery>=3.27"]
snowflake = ["snowflake-connector-python>=3.12"]
postgres = ["psycopg[binary]>=3.2"]
all = [
    "google-cloud-bigquery>=3.27",
    "snowflake-connector-python>=3.12",
    "psycopg[binary]>=3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alma_connectors"]
